Added cart fragment nested recyclerviews

+++ app/src/main/java/edu/curtin/foodapp/model/orders/OrderItem.java
@@ -1,13 +1,11 @@
 package edu.curtin.foodapp.model.orders;
 
-
-
 public class OrderItem {
     String name;
-    float price;
+    double price;
     int quantity;
 
-    public OrderItem(String name, float price, int quantity) {
+    public OrderItem(String name, double price, int quantity) {
         this.name = name;
         this.price = price;
         this.quantity = quantity;
@@ -23,13 +21,13 @@ public class OrderItem {
         this.name = name;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
     // Get total price of the item with quantity
-    public float getTotalPrice() {
-        float price = this.price;
+    public double getTotalPrice() {
+        double price = this.price;
         int quantity = this.quantity;
         return price * (float) quantity;
     }
